requirement.txt
Django
PyPDF2
python-docx
reportlab